Move app/config/app to config/app
