+++ app/server.jsx
@@ -4,7 +4,7 @@ import createRoutes from './routes';
 import configureStore from './store/configureStore';
 import * as types from './types';
 import preRenderMiddleware from './middlewares/preRenderMiddleware';
-import { host, port } from './config/app';
+import { host, port } from '../config/app';
 import pageRenderer from './utils/pageRenderer';
 
 // configure baseURL for axios requests (for serverside API calls)
